refactor: rework internals and external api
This is a large rework of how the d2-style tool works in order to
simplify usage and allow for more flexible usage.

Notable changes:

- Removes lint-staged and removes automatically staging formatted files.

- Only runs checks when commiting by default to ensure consistency.

- Includes recommended ESLint rules in the default ESLint config.

- Includes a React project.

- Renames `setup` to `install` and adds a `--list-groups` option.
`install` is preferred to be able to pair it with an `uninstall` command
in the future.

- Adds command to use Prettier to format structured text files (html,
yaml, markdown, etc.).

- Uses `npx` internally to execute command binaries instead of relying
on the NodeJS API, this makes the entire code base a lot simpler to work
with and to add new commands.

- `check` and `apply` commands now accept a `--pattern` to match files.

- Uses `fast-glob` to grab all files that match a pattern, then filters
them against other parameters, e.g. if they are staged or if only a
single file is given.

BREAKING CHANGE: Removes LintStaged, renames `setup` to `install`,
removes `validate` command. Uses a different way to match files and
filters out results instead.