fix: add CHANGELOG.md to blacklist
When the semantic-release plugin generates the CHANGELOG.md file and
tries to commit it, it runs into the git commit hook that validates the
format of markdown files, and it fails.

The proper fix would probably be to apply the code-style for the
CHANGELOG.md file between the point where it is generated and it is
commited, but that requires changes in the `release` command of d2, and
since it is based around semantic-release we need to figure out how to
work around it.

The easiest is just to ignore formatting on the CHANGELOG.md file for
now.