fix(commands): run commands with yarn
Use yarn over npx to execute commands, since we use Yarn collectively.

This should solve an issue where the tools cannot be found[1] when
d2 is installed globally with Yarn, and d2-style uses npx to execute the
tools internally.

The dependencies are installed to where Yarn keeps deps, but npx would
look for them in a different place.

[1] https://github.com/dhis2/cli-style/issues/216