feat(install): prompt to choose project from list
When running `d2 style install` without arguments, an interative prompt
displaying the available projects is displayed:

```
$ d2-style install
? Choose the project template (Use arrow keys)
❯ base
  js
  react
```

To install a custom set of groups, it works the same as before. The
groups are not displayed in the prompt to guide the user to make a safe
choice.

The full list can be seen with `--list-groups` and multiple groups can
be passed in as a space separated list as before.

This feature is kept small and simple in scope to avoid the temptation
to refactor how projects and groups function and tackle that as a
separate task.