fix: update bin lookup to avoid assuming node_modules lookup
This works in both yarn1 and yarn2 environments, and should be more robust against (for instance) transitive dependencies which don't get their executables hoisted to the top-level `node_modules/.bin` (which is the case in yarn2 - transitive dependencies aren't available with `yarn run`)

The main downside here is that it assumes the `bin` file is a `node` script - this is also the behavior of yarn's own `run` command, so it _should_ be OK - we're also only using Node scripts from dependencies in `cli-stye` so far, so it should be fine for now.  The challenge with supporting non-node scripts is that they can't be run directly from within a .zip file (where they exist in yarn2), so we need to run `yarn node` to be able to resolve plug-and-play files and dependencies.