feat: warnings in eslint
Allow d2-style to print warnings from ESLint.

Set a zero tolerance on warnings by making ESLint exit with a non-zero
code if there are any warnings.

A developer is required to either:

a)  fix the warning
b)  explicitly mark the code that triggered the warning with an
eslint-disable comment.

A good disable directive uses the comment syntax for ESLint:

    /* eslint-disable [rule] -- [comment] */

E.g.

    /* eslint-disable-next-line no-console -- we need to log here */