refactor: drop eslint plugin for prettier
First of all, if an editor is configured to run ESLint and show
warnings/errors in the code, then all the mistakes that Prettier will
automatically fix are shoved in the user's face, which is the opposite
from what Prettier is intended to do.

Second of all, the Prettier binary is much faster than the Prettier
ESLint plugin.