feat: new and improved d2-style (#378)
"d2-style" no longer tries to obfuscate what tools are used under the hood, and
leans into being a friendly helper to nudge developers to comply with the dhis2
code style and guidelines.

New features:

-   "d2-style <check/apply>" can now run all configured linters and formatters
in a single command, making it easier to use.

-   "d2-style <add/remove>" is used to add configuration files for individual
tools, e.g. eslint, prettier, etc.

-   "d2-style add" supports templates for e.g. react using:

		d2-style add eslint react

-   "d2-style install" reads the project configuration from .d2/style.config.js
and sets up the project in accordance to that.

BREAKING CHANGE: The verb (check/apply) is now moved to the top-level.  E.g.
"d2-style js check" becomes "d2-style check js". This is to allow all checkers
to run with a single command: "d2-style check"

BREAKING CHANGE: "d2-style install" is no longer used to set up linters.  As of
husky@5 the tool is vastly simplified and much faster. Hooks can be installed
manually with "d2-style install", but is also run as a "post-install" script
that we control for consistency. Configuration is added to the project with the
"d2-style add" command.