fix: do not run tools without configs
If configuration files for e.g. Prettier and/or ESLint are missing, they
will still be run if we can find files in the project that should be
checked, except for ls-lint which we skip if it's not configured.

This makes it so that we only run tools when there is a local
configuration (i.e. the tool is installed in the project).